Swift Temporal SDK code
This commit contains the entire contents of the Swift Temporal SDK. At a high level, it provides a TemporalClient and a TemporalWorker. The client is used to interact with the Temporal APIs. The worker allows registering of activities and workflows. The Swift Temporal SDK integrates natively with Swift Concurrency when authoring activities and workflows.

Co-authored-by: Daniel Grumberg <dgrumberg@apple.com>
Co-authored-by: Philipp Zagar <p_zagar@apple.com>
Co-authored-by: Rob Leith <robert_leith@apple.com>
Co-authored-by: Gabbi Fisher <gabriele.fisher@apple.com>
Co-authored-by: Kesi Maduka <kezi@apple.com>
Co-authored-by: Dominic Morneau <morneau@apple.com>
Co-authored-by: Barry Tomasini <tomasini@apple.com>
Co-authored-by: Andreas Bauer <andreas-bauer@apple.com>
Co-authored-by: Michael Gecht <gecht@apple.com>
Co-authored-by: Joe Heck <j_heck@apple.com>
Co-authored-by: Morris Cornell-Morgan <mcornellmorgan@apple.com>