Add line on how to install/upgrade Swift to README (#8)
### Motivation

Since right now the vast majority of the Temporal community uses
something _other_ than Swift :) it might be nice to throw them a
breadcrumb on how to get it.

In my case I had it, but the wrong version (6.1.0) and because I wasn't
familiar with Swift (yet!) I wasn't sure how to proceed. Hoping to make
this easier for the next person.

### Modifications

One liner to add a pointer to https://www.swift.org/install/ as per
Temporal Community Slack.

### Result

^ :)

### Test Plan

N/A

Co-authored-by: Franz Busch <f.busch@apple.com>