Activity Execution Info to support checking if heartbeat details from previous attempt are available (#41)
### Motivation

Currently, we offer `heartbeatDetails(as:)` method to retrieve the
activity details of an activity execution. If it's the first attempt,
this method fails with an ArgumentError `Mismatched number of values and
payloads` as there are no heartbeat details available. The user has no
way to differentiate between no heartbeat details being available or
that they did indeed supple the wrong amount of payload types.

### Modifications

We have `heartbeatDetails(as:)` return an optional tuple instead, so
users can check if there are heartbeat details available from previous
activity attempts.

### Test Plan

Augmented existing unit tests to check activity details to be present or
not.